+++ app/Http/Controllers/RoomController.php
@@ -152,7 +152,6 @@ class RoomController extends Controller
                 unlink('storage/'.$room->room_picture);
             }
 
-            
             if($user->id == $room->hotel->user_id && $user->user_level == 1){
                 $file = $request->file('room_picture');
                 $upload_dest = 'room_picture';
@@ -185,7 +184,9 @@ class RoomController extends Controller
                 if($validator->fails()){
                     return response()->json($validator->errors()->toJson(), 400);
                 }
-                unlink('storage/'.$room->room_picture);
+                if($room->room_picture != null){
+                    unlink('storage/'.$room->room_picture);
+                }
                 $file = $request->file('room_picture');
                 $upload_dest = 'room_picture';
                 $extension = $file->extension();
